requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
DateTime==5.5
et-xmlfile==1.1.0
idna==3.6
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
setuptools==69.5.1
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
zope.interface==6.3
